tests/wpt/web-platform-tests/navigation-timing/nav2_test_open_blank_page.html

<!DOCTYPE html>
<html>
    <head>
        <meta charset="utf-8" />
        <title>Navigation Timing 2 WPT</title>
        <link rel="author" title="Google" href="http://www.google.com/" />
        <link rel="help" href="http://www.w3.org/TR/navigation-timing/#sec-navigation-timing-interface"/>
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>

        <script>

            function onload_test()
            {
                var iframe_performance = document.getElementById("frameContext").contentWindow.performance;
                assert_equals(iframe_performance.getEntriesByType("navigation").length, 0, "Expected there is no navigation timing instance");
                done();
            }
        </script>

    </head>
    <body>
        <h1>Description</h1>
        <p>This test validates there does not exist a navigation timing instance when opening internal pages like about:blank.</p>

        <iframe id="frameContext" onload="onload_test();" src="about:blank" style="width: 250px; height: 250px;"></iframe>
    </body>
</html>
